05 Validation | response Record not found!

+++ app/Http/Controllers/Country/CountryController.php
@@ -15,6 +15,11 @@ class CountryController extends Controller
 
     public function countryByID($id)
     {
+        $country = CountryModel::find($id);
+        if(is_null($country)){
+
+            return response()->json("Record not found!",404);
+        }
         return response()->json(CountryModel::find($id),200);
     }
 
@@ -24,13 +29,21 @@ class CountryController extends Controller
         return response()->json($country,201);
     }
 
-    public function countryUpdate(Request $request, CountryModel $country)
+    public function countryUpdate(Request $request, $id)
     {
+        $country = CountryModel::find($id);
+        if(is_null($country)){
+            return response()->json("Record not found!", 404);
+        }
         $country->update($request->all());
         return response()->json($country,200);
     }
-    public function countryDelete(Request $request, CountryModel $country)
+    public function countryDelete(Request $request, $id)
     {
+        $country = CountryModel::find($id);
+        if (is_null($country)) {
+            return response()->json("Record not found!", 404);
+        }
         $country->delete();
         return response()->json(null,204);
     }
